Data.Certificate.X509

Description

Read/Write X509 certificate

Data Structure

data X509

Constructors

X509

Fields

x509Cert :: Certificate

the certificate part of a X509 structure

x509CachedSigningData :: Maybe ByteString

a cache of the raw representation of the x509 part for signing since encoding+decoding might not result in the same data being signed.

x509CachedData :: Maybe ByteString

a cache of the raw representation of the whole x509.

x509SignatureALG :: SignatureALG

the signature algorithm used.

x509Signature :: [Word8]

the signature.

Data Structure (reexported from X509Cert)

data SignatureALG

Constructors

SignatureALG HashALG PubKeyALG

SignatureALG_Unknown OID

data HashALG

Constructors

HashMD2

HashMD5

HashSHA1

HashSHA224

HashSHA256

HashSHA384

HashSHA512

data PubKeyALG

Constructors

PubKeyALG_RSA

PubKeyALG_DSA

PubKeyALG_ECDSA

PubKeyALG_DH

PubKeyALG_Unknown OID

data PubKey

Constructors

PubKeyRSA PublicKey
RSA public key

PubKeyDSA PublicKey
DSA public key

PubKeyDH (Integer, Integer, Integer, Maybe Integer, ([Word8], Integer))
DH format with (p,g,q,j,(seed,pgenCounter))

PubKeyECDSA ECDSA_Hash ByteString
ECDSA format not done yet FIXME

PubKeyUnknown OID [Word8]
unrecognized format

type OID = [Integer]

data ASN1StringType

Constructors

UTF8

Printable

Univ

BMP

IA5

T61

type ASN1String = (ASN1StringType, String)

newtype DistinguishedName

Constructors

DistinguishedName

Fields

getDistinguishedElements :: [(OID, ASN1String)]

data Certificate

Constructors

Certificate

Fields

certVersion :: Int

Certificate Version

certSerial :: Integer

Certificate Serial number

certSignatureAlg :: SignatureALG

Certificate Signature algorithm

certIssuerDN :: DistinguishedName

Certificate Issuer DN

certSubjectDN :: DistinguishedName

Certificate Subject DN

certValidity :: (Time, Time)

Certificate Validity period

certPubKey :: PubKey

Certificate Public key

certExtensions :: Maybe [ExtensionRaw]

Certificate Extensions

module Data.Certificate.X509.Ext

helper for signing/veryfing certificate

getSigningData :: X509 -> ByteString

get signing data related to a X509 message, - which is either the cached data or the encoded certificate

serialization from ASN1 bytestring

decodeCertificate :: ByteString -> Either String X509

decode an X509 from a bytestring - the structure is the following: - Certificate - Certificate Signature Algorithm - Certificate Signature

encodeCertificate :: X509 -> ByteString

encode a X509 certificate to a bytestring

Distinguished names related function

decodeDN :: ByteString -> Either String DistinguishedName

encodeDN :: DistinguishedName -> ByteString

hashDN :: DistinguishedName -> ByteString

Make an openssl style hash of distinguished name

hashDN_old :: DistinguishedName -> ByteString

Create an openssl style old hash of distinguished name